# ownerpay/openapi.json
{"openapi":"3.1.0","info":{"title":"OwnerPay API","version":"0.1.0"},"paths":{"/rc/estimate":{"post":{"summary":"Estimate","operationId":"estimate_rc_estimate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstimateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstimateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rc/memo/{id}":{"get":{"summary":"Memo","operationId":"memo_rc_memo__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Business":{"properties":{"entity_type":{"type":"string","pattern":"^(s_corp)$","title":"Entity Type"},"revenue":{"type":"number","minimum":0.0,"title":"Revenue"},"profit":{"type":"number","title":"Profit"},"headcount":{"type":"integer","minimum":0.0,"title":"Headcount"}},"type":"object","required":["entity_type","revenue","profit","headcount"],"title":"Business"},"EstimateRequest":{"properties":{"tax_year":{"type":"integer","maximum":2026.0,"minimum":2024.0,"title":"Tax Year"},"location":{"$ref":"#/components/schemas/Location"},"business":{"$ref":"#/components/schemas/Business"},"owner":{"$ref":"#/components/schemas/Owner"},"role_split":{"items":{"$ref":"#/components/schemas/RoleSplit"},"type":"array","title":"Role Split"}},"type":"object","required":["tax_year","location","business","owner","role_split"],"title":"EstimateRequest"},"EstimateResponse":{"properties":{"id":{"type":"string","title":"Id"},"low":{"type":"number","title":"Low"},"median":{"type":"number","title":"Median"},"high":{"type":"number","title":"High"},"recommended":{"type":"number","title":"Recommended"},"soc_sources":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Soc Sources","default":[]},"adjustments":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Adjustments","default":[]},"flags":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Flags","default":[]},"assumptions":{"items":{"type":"string"},"type":"array","title":"Assumptions","default":[]},"memo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo Url"}},"type":"object","required":["id","low","median","high","recommended"],"title":"EstimateResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Location":{"properties":{"state":{"type":"string","pattern":"^[A-Z]{2}$","title":"State"},"msa_code":{"type":"string","pattern":"^\\d{5}$","title":"Msa Code"}},"type":"object","required":["state","msa_code"],"title":"Location"},"MemoResponse":{"properties":{"id":{"type":"string","title":"Id"},"html_url":{"type":"string","title":"Html Url"},"pdf_url":{"type":"string","title":"Pdf Url"}},"type":"object","required":["id","html_url","pdf_url"],"title":"MemoResponse"},"Owner":{"properties":{"hours_per_week":{"type":"integer","maximum":80.0,"minimum":0.0,"title":"Hours Per Week"},"experience_years":{"type":"integer","maximum":60.0,"minimum":0.0,"title":"Experience Years"}},"type":"object","required":["hours_per_week","experience_years"],"title":"Owner"},"RoleSplit":{"properties":{"role":{"type":"string","pattern":"^(sales|operations|administration|finance|marketing|technical)$","title":"Role"},"percent":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Percent"}},"type":"object","required":["role","percent"],"title":"RoleSplit"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"servers":[{"url":"https://api-staging.ownerpay.dev"},{"url":"https://api.ownerpay.dev"},{"url":"http://localhost:8000"}]}